#152: algorithm renaming and removing generics
